# example-project/cmd/web-api/docs/swagger.yaml
basePath: '{{.BasePath}}'
definitions:
...
  user.Token:
    properties:
      access_token:
        type: string
      expiry:
        type: string
      token_type:
        type: string
    type: object
  user.UserArchiveRequest:
    properties:
      id:
        example: d69bdef7-173f-4d29-b52c-3edc60baf6a2
        type: string
    required:
    - id
    type: object
  user.UserCreateRequest:
    properties:
      email:
        example: gabi@geeksinthewoods.com
        type: string
      name:
        example: Gabi May
        type: string
      password:
        example: SecretString
        type: string
      password_confirm:
        example: SecretString
        type: string
      timezone:
        example: America/Anchorage
        type: string
    required:
    - email
    - name
    - password
    type: object
  user.UserResponse:
    properties:
      archived_at:
        $ref: '#/definitions/web.TimeResponse'
        type: object
      created_at:
        $ref: '#/definitions/web.TimeResponse'
        type: object
      email:
        example: gabi@geeksinthewoods.com
        type: string
      id:
        example: d69bdef7-173f-4d29-b52c-3edc60baf6a2
        type: string
      name:
        example: Gabi May
        type: string
      timezone:
        example: America/Anchorage
        type: string
      updated_at:
        $ref: '#/definitions/web.TimeResponse'
        type: object
    type: object
  user.UserUpdatePasswordRequest:
    properties:
      id:
        example: d69bdef7-173f-4d29-b52c-3edc60baf6a2
        type: string
      password:
        example: NeverTellSecret
        type: string
      password_confirm:
        example: NeverTellSecret
        type: string
    required:
    - id
    - password
    type: object
  user.UserUpdateRequest:
    properties:
      email:
        example: gabi.may@geeksinthewoods.com
        type: string
      id:
        example: d69bdef7-173f-4d29-b52c-3edc60baf6a2
        type: string
      name:
        example: Gabi May Not
        type: string
      timezone:
        example: America/Anchorage
        type: string
    required:
    - id
    type: object
...
  web.Error:
    properties:
      err:
        type: error
      fields:
        items:
          $ref: '#/definitions/web.FieldError'
        type: array
      status:
        type: integer
    type: object
  web.ErrorResponse:
    properties:
      error:
        type: string
      fields:
        items:
          $ref: '#/definitions/web.FieldError'
        type: array
    type: object
  web.FieldError:
    properties:
      error:
        type: string
      field:
        type: string
    type: object
  web.TimeResponse:
    properties:
      date:
        example: "2019-06-25"
        type: string
      kitchen:
        example: 3:00AM
        type: string
      local:
        example: Tue Jun 25 3:00AM
        type: string
      local_date:
        example: Tue Jun 25
        type: string
      now_rel_time:
        example: 15 hours from now
        type: string
      now_time:
        example: 5 hours ago
        type: string
      rfc1123:
        example: Tue, 25 Jun 2019 03:00:53 AKDT
        type: string
      time:
        example: "03:00:53"
        type: string
      timezone:
        example: America/Anchorage
        type: string
      value:
        example: "2019-06-25T03:00:53.284-08:00"
        type: string
      value_utc:
        example: "2019-06-25T11:00:53.284Z"
        type: string
    type: object
host: '{{.Host}}'
info:
  contact:
    email: support@geeksinthewoods.com
    name: API Support
    url: http://example.com/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://example.com/terms
  title: SaaS Example API
  version: '{{.Version}}'
paths:
...
  /oauth/token:
    post:
      consumes:
      - application/json
      description: Token generates an oauth2 accessToken using Basic Auth with a user's
        email and password.
      parameters:
      - description: Scope
        enum:
        - user
        - admin
        in: query
        name: scope
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            $ref: '#/definitions/user.Token'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.Error'
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web.Error'
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.Error'
            type: object
      security:
      - BasicAuth: []
      summary: Token handles a request to authenticate a user.
      tags:
      - user
...
  /users:
    get:
      consumes:
      - application/json
      description: Find returns the existing users in the system.
      parameters:
      - description: 'Filter string, example: name = ''Company Name'' and email =
          ''gabi.may@geeksinthewoods.com'''
        in: query
        name: where
        type: string
      - description: 'Order columns separated by comma, example: created_at desc'
        in: query
        name: order
        type: string
      - description: 'Limit, example: 10'
        in: query
        name: limit
        type: integer
      - description: 'Offset, example: 20'
        in: query
        name: offset
        type: integer
      - description: 'Included Archived, example: false'
        in: query
        name: included-archived
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/user.UserResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
      security:
      - OAuth2Password: []
      summary: List users
      tags:
      - user
    patch:
      consumes:
      - application/json
      description: Update updates the specified user in the system.
      parameters:
      - description: Update fields
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/user.UserUpdateRequest'
          type: object
      produces:
      - application/json
      responses:
        "201": {}
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
      security:
      - OAuth2Password: []
      summary: Update user by ID
      tags:
      - user
    post:
      consumes:
      - application/json
      description: Create inserts a new user into the system.
      parameters:
      - description: User details
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/user.UserCreateRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
      security:
      - OAuth2Password: []
      summary: Create new user.
      tags:
      - user
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete removes the specified user from the system.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201": {}
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
      security:
      - OAuth2Password: []
      summary: Delete user by ID
      tags:
      - user
    get:
      consumes:
      - application/json
      description: Read returns the specified user from the system.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
      security:
      - OAuth2Password: []
      summary: Get user by ID
      tags:
      - user
  /users/archive:
    patch:
      consumes:
      - application/json
      description: Archive soft-deletes the specified user from the system.
      parameters:
      - description: Update fields
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/user.UserArchiveRequest'
          type: object
      produces:
      - application/json
      responses:
        "201": {}
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
      security:
      - OAuth2Password: []
      summary: Archive user by ID
      tags:
      - user
  /users/password:
    patch:
      consumes:
      - application/json
      description: Update updates the password for a specified user in the system.
      parameters:
      - description: Update fields
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/user.UserUpdatePasswordRequest'
          type: object
      produces:
      - application/json
      responses:
        "201": {}
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
      security:
      - OAuth2Password: []
      summary: Update user password by ID
      tags:
      - user
  /users/switch-account/{account_id}:
    patch:
      consumes:
      - application/json
      description: SwitchAccount updates the auth claims to a new account.
      parameters:
      - description: Account ID
        in: path
        name: account_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201": {}
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
      security:
      - OAuth2Password: []
      summary: Switch account.
      tags:
      - user
securityDefinitions:
  BasicAuth:
    type: basic
  OAuth2Password:
    flow: password
    scopes:
      admin: ' Grants administrative privileges with role of admin.'
      user: ' Grants basic privileges with role of user.'
    tokenUrl: /v1/oauth/token
    type: oauth2
swagger: "2.0"